requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
pydantic
python-multipart
jinja2
aiofiles
python-jose
passlib
bcrypt